:bug: Process previous effect, not event
Fixes #105

Previously, the previous event would be saved. This caused multiple issues where additonal hooks for the event were processed, so things like cooling tracking would be restarted when heating progress was finished, if that was the last event to come through.

Now, it records the effect, rather than event that should be run, and bypasses the additonal event handling. Much better...